[Bugfix]: Fix Event modal submit button disable state (#255)
* separate modal content into separate components, fix button disable state

* remove unused styles
